moved TypedData to a struct in EthereumTypedMessage and used it to replace Messsage<TypedData>, hence retiring Message.java. SignedMessage.java is also deleted because no one uses it.
