+++ app/src/main/java/com/alphawallet/app/web3/OnSetValuesListener.java
@@ -1,7 +1,5 @@
 package com.alphawallet.app.web3;
 
-import com.alphawallet.app.web3.entity.Message;
-
 import java.util.Map;
 
 /**
