+++ app/src/main/java/com/alphawallet/app/web3/SignCallbackJSInterface.java
@@ -6,10 +6,9 @@ import android.webkit.JavascriptInterface;
 import android.webkit.WebView;
 
 import com.alphawallet.token.entity.EthereumMessage;
+import com.alphawallet.token.entity.EthereumTypedMessage;
 import com.google.gson.Gson;
 import com.alphawallet.app.web3.entity.Address;
-import com.alphawallet.app.web3.entity.Message;
-import com.alphawallet.app.web3.entity.TypedData;
 import com.alphawallet.app.web3.entity.Web3Transaction;
 
 import java.math.BigInteger;
@@ -80,11 +79,11 @@ public class SignCallbackJSInterface {
         webView.post(() -> {
             TrustProviderTypedData[] rawData = new Gson().fromJson(data, TrustProviderTypedData[].class);
             int len = rawData.length;
-            TypedData[] typedData = new TypedData[len];
+            EthereumTypedMessage.TypedData[] typedData = new EthereumTypedMessage.TypedData[len];
             for (int i = 0; i < len; i++) {
-                typedData[i] = new TypedData(rawData[i].name, rawData[i].type, rawData[i].value);
+                typedData[i] = new EthereumTypedMessage.TypedData(rawData[i].name, rawData[i].type, rawData[i].value);
             }
-            onSignTypedMessageListener.onSignTypedMessage(new Message<>(typedData, getUrl(), callbackId));
+            onSignTypedMessageListener.onSignTypedMessage(new EthereumTypedMessage(typedData, getUrl(), callbackId));
         });
     }
